pyproject.toml
[tool.poetry]
name = "m"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.1"
requests = "^2.28.2"
sqlalchemy = "^2.0.10"
alembic = "^1.10.3"
mysql-connector-python = "^8.0.33"
python-dotenv = "^1.0.0"
tenacity = "^8.2.2"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
